+++ CHANGELOG.md
@@ -14,5 +14,6 @@ This changelog only contains the changes that are unreleased. For changes for in
 - Unable to export when name contains invalid path characters
 - Remove overrides folder from exports when it's empty
 - Folders to go into overrides not working as intended when exporting an instance
+- Exclude .fabric and .quilt folders from being exported
 
 ### Misc